+++ src/backend/marsha/core/tests/test_models_video.py
@@ -1,5 +1,8 @@
 """Tests for the models in the ``core`` app of the Marsha project."""
+from django.core.exceptions import ValidationError
 from django.test import TestCase
+
+from safedelete.models import SOFT_DELETE_CASCADE
 
 from ..factories import VideoFactory
 
@@ -18,3 +21,49 @@ class VideoModelsTestCase(TestCase):
 
         video.delete()
         self.assertEqual(str(video), "j'espère [deleted]")
+
+    def test_models_video_fields_lti_id_unique(self):
+        """Videos should be unique for a given duo: lti_id/playlist."""
+        video = VideoFactory()
+
+        # A video with a different lti_id and the same playlist can still be created
+        VideoFactory(playlist=video.playlist)
+
+        # A video for a different playlist and the same lti_id can still be created
+        VideoFactory(lti_id=video.lti_id)
+
+        # Trying to create a video with the same duo lti_id/playlist should raise a
+        # database error
+        with self.assertRaises(ValidationError) as context:
+            VideoFactory(lti_id=video.lti_id, playlist=video.playlist)
+        self.assertEqual(
+            context.exception.messages,
+            ["Video with this Lti id and Playlist already exists."],
+        )
+
+        # Soft deleted videos should not count for unicity
+        video.delete(force_policy=SOFT_DELETE_CASCADE)
+        VideoFactory(lti_id=video.lti_id, playlist=video.playlist)
+
+    def test_models_video_fields_resource_id_unique(self):
+        """Videos should be unique for a given duo: resource_id/playlist."""
+        video = VideoFactory()
+
+        # A video with a different resource_id and the same playlist can still be created
+        VideoFactory(playlist=video.playlist)
+
+        # A video for a different playlist and the same resource_id can still be created
+        VideoFactory(resource_id=video.resource_id)
+
+        # Trying to create a video with the same duo resource_id/playlist should raise a
+        # database error
+        with self.assertRaises(ValidationError) as context:
+            VideoFactory(resource_id=video.resource_id, playlist=video.playlist)
+        self.assertEqual(
+            context.exception.messages,
+            ["Video with this Resource UUID and Playlist already exists."],
+        )
+
+        # Soft deleted videos should not count for unicity
+        video.delete(force_policy=SOFT_DELETE_CASCADE)
+        VideoFactory(resource_id=video.resource_id, playlist=video.playlist)